feat(#1683): substitui CalculadoraPeriodosAvaliativos por GeradorPeriodosAvaliativos
- Remove CalculadoraPeriodosAvaliativos (Eloquent direto, sem repository)
- GeradorPeriodosAvaliativos usa PlanoTrabalhoConsolidacaoRepository
- Testa todas as periodicidades: MENSAL, BIMESTRAL, TRIMESTRAL, SEMESTRAL, DIAS, SEMANAL, QUINZENAL
- Edge cases: fevereiro bissexto/não-bissexto, jul-ago, último período truncado, dia único, sem gaps

+++ back-end/app/V2/PlanoTrabalho/Consolidacao/GeradorPeriodosAvaliativos.php
@@ -47,7 +47,7 @@ class GeradorPeriodosAvaliativos
 
     private function calcularProximaData(string $data, Programa $programa): string
     {
-        $dayWeek = date('w', strtotime($data));
+        $dayWeek = (int) date('w', strtotime($data));
 
         return match ($programa->periodicidade_consolidacao) {
             'DIAS' => date('Y-m-d', strtotime($data . ' + ' . $programa->periodicidade_valor . ' days')),
